UWORD to FourCC ID
Naughty workaround for IFF chunk headers, which require a printable type for the Class Type. Parrot bends the rule a little, and prefers the asset ID to be stored in the class type field. Passing a UWORD id in, causes IFFLibrary to fail due to internal checks.

Instead each nibble of the UWORD is spread across the four bytes with an offset of 0x41 "A", so the number is presented as printable characters.