Refinement on the DIALOGUE_TEXT design.
Now has designs for DIALOGUE pointers which can share the same 4 bytes as a DIALOGUE_REF in a struct.

Dialogue References as a ULONG are in the range of 0x00FC0000 to 0x00FCFFFFF which is a reserved memory address used by the Amiga Kickstart. String pointers will never be that.